fix(pwa): keep the public header on errors thrown in public pages
Review on #764: the root error.tsx hard-coded the app chrome, so an error
thrown while browsing a public page (FAQ/legal/privacy) flashed the
authenticated TopBar. Extract the error card into <ErrorBoundaryContent> and add
(public)/error.tsx wrapping it with <SiteChrome variant="public">.

not-found is intentionally left at the root only: arbitrary 404s hit the root
not-found regardless (route groups create no URL segment / catch-all) and the
public pages never call notFound(), so a (public)/not-found would be dead code.

Refs #649
